LayoutTests/platform/glib/imported/w3c/web-platform-tests/focus/focus-restoration-in-different-site-iframes-window-expected.txt


FAIL Check result assert_equals: Check log expected "outerlog:log:willfocuswindow,windowfocus,didfocuswindow,windowblur,windowfocus," but got "outerlog:log:willfocuswindow,didfocuswindow,"
